+++ app/build.gradle
@@ -3,7 +3,7 @@ apply plugin: 'com.android.application'
 android {
     compileSdkVersion 27
     defaultConfig {
-        applicationId "mashup.kr.mapc"
+        applicationId "kr.mashup.mapc"
         minSdkVersion 24
         targetSdkVersion 27
         versionCode 1
