+++ app/src/main/java/kr/mashup/mapc/TypeFaceUtil.java
@@ -1,4 +1,4 @@
-package mashup.kr.mapc;
+package kr.mashup.mapc;
 
 import android.content.Context;
 import android.graphics.Typeface;
